django_oauth/unicorn/api/serializer.py
from rest_framework import serializers
from unicorn.models import Unicorn


class UnicornSerializer(serializers.ModelSerializer):
    class Meta:
        model = Unicorn
        fields = ('id', 'name', 'age')
