django_oauth/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
mysqlclient
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
django-cors-headers==3.7.0
django-crispy-forms==1.11.0
django-extra-views==0.13.0
django-social-auth==0.7.28
django-cors-middleware==1.5.0
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
httplib2==0.19.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.0
MarkupSafe==2.0.0
oauth2==1.9.0.post1
oauthlib==3.1.0
packaging==20.9
Pillow==8.1.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python-openid==2.2.5
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4